optimizations and fixes

+++ train-main.sh
@@ -26,9 +26,13 @@ fi
 
 echo "Downloading dataset to: ${DATASET_PATH}"
 mkdir -p "${DATASET_PATH}" || true
-curl -L -o "${DATASET_PATH}/dataset.zip" \
-    "${DATASET_HOST}/${DATASET_NAME}"
-(cd "${DATASET_PATH}" && unzip "${DATASET_PATH}/dataset.zip")
+
+if [ ! -f "${DATASET_PATH}/dataset.zip" ];
+then
+    curl -L -o "${DATASET_PATH}/dataset.zip" \
+        "${DATASET_HOST}/${DATASET_NAME}"
+    (cd "${DATASET_PATH}" && unzip "${DATASET_PATH}/dataset.zip")
+fi
 
 BASE_MODEL="${BASE_MODEL:-flux}"
 echo "Running ${BASE_MODEL} training..."
